pageView bug fix on staging environment. Refactor URL validation in post API
Previously, a check was made to ensure that the URL provided started with "https://www.gitpod.io", which is adjusted with additional check to validate URLs containing "gitpod-kumquat.netlify.app". If the URL is invalid, the code returns a 400 status code and error message, and continues with the request if the URL is valid. This commit fixes the pageView bug on the staging environment. Occured in #3369